Move the eigenvalue routines to genmap-eigen.c
